feat(components): add logout button to Nav component
In this commit, a logout button was added to the Nav component. The button appears when the user is logged in, and it allows the user to log out of the application. If the user is not logged in, the button will not appear.  To implement this feature, a check for the presence of a token was added, and if the token exists, the logout button is displayed. When the user clicks the logout button, they are logged out of the application.